refactor: move types.ts file into types folder & update alias
